refactor: move Lotto class
